+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Ability to Block-list plugins that should not be installed on the jenkins dogu
+  - Block-listed plugins will be automatically removed when found to be installed on dogu start and restart
 
 ### Changed
 - Upgrade Jenkins to 2.462.2; #171